Fixed some things, rearranged mnist with better project structure
