+++ test/mnist/Dockerfile
@@ -1,3 +1,3 @@
-FROM scaleoutsystems/fedn-client:develop
+FROM scaleoutsystems/fedn-combiner:develop
 COPY fedn-network.yaml /app/ 
 WORKDIR /app
