+++ test/mnist/client/models/mnist_model.py
@@ -1,0 +1,34 @@
+import keras
+from keras.layers import Conv2D, MaxPooling2D
+from keras.layers import Dense, Dropout, Flatten
+from keras.models import Sequential
+import os
+
+
+import tempfile
+
+# Create an initial CNN Model
+def create_seed_model():
+	# input image dimensions
+	img_rows, img_cols = 28, 28
+	input_shape = (img_rows, img_cols, 1)
+	num_classes = 10
+
+	model = Sequential()
+	model.add(Conv2D(32, kernel_size=(3, 3),
+                 activation='relu',
+                 input_shape=input_shape))
+
+	model.add(Conv2D(64, (3, 3), activation='relu'))
+	model.add(MaxPooling2D(pool_size=(2, 2)))
+	model.add(Dropout(0.25))
+	model.add(Flatten())
+	model.add(Dense(128, activation='relu'))
+	model.add(Dropout(0.5))
+	model.add(Dense(num_classes, activation='softmax'))
+
+	model.compile(loss=keras.losses.categorical_crossentropy,
+    	          optimizer=keras.optimizers.Adadelta(),
+        	      metrics=['accuracy'])
+	return model
+
