+++ test/mnist/client/validate.py
@@ -3,7 +3,7 @@ import tensorflow as tf
 tf.compat.v1.logging.set_verbosity(tf.compat.v1.logging.ERROR)
 import tensorflow.keras as keras
 import tensorflow.keras.models as krm
-from read_data import read_data
+from data.read_data import read_data
 import pickle
 import json
 from sklearn import metrics
@@ -101,7 +101,7 @@ if __name__ == '__main__':
     helper = KerasWeightsHelper()
     weights = helper.load_model(sys.argv[1])
 
-    from keras_model_structure import create_seed_model
+    from models.mnist_model import create_seed_model
 
     model = create_seed_model()
     model.set_weights(weights)
