+++ test/mnist/docker-compose.dev.yaml
@@ -3,7 +3,7 @@ services:
   client:
     environment:
       - GET_HOSTS_FROM=dns
-    image: "scaleoutsystems/mnist-client:latest"
+    image: "scaleoutsystems/mnist-client:develop"
     build:
       context: .
     working_dir: /app
